+++ xtract/xtract_types.h
@@ -33,12 +33,12 @@ extern "C" {
 /* \brief Data structure used to store amplitude data between calls to xtract_attack_time and other functions. */
 typedef struct _xtract_amp_tracker {
     int count;
-    float previous_amp;
+    double previous_amp;
 } xtract_amp_tracker;
 
 typedef struct _xtract_frame_tracker {
     int frame_count;
-    float *previous_frame;
+    double *previous_frame;
 } xtract_frame_tracker;
 
 #ifdef __cplusplus
